buttons are working
the forward and back button will count and subtract and loop back around the set lenth. The current issue is how to grab the info i need from the other files and use it. The bigger issue is how to define a index of the array i want to display. That would be linked to the button count.